finish and reformatting
