+++ src/components/__init__.py
@@ -2,3 +2,6 @@
 
 from .bar_chart import bar_chart
 from .map import park_map, create_markers
+from .filter import neighbourhood_dropdown, facility_dropdown, special_feature_dropdown, washrooms_checkbox
+from .cards import avg_hectare_card, num_parks_card
+from .modal import park_info_modal
